Bonus

+++ src/main/java/github/Issue.java
@@ -46,9 +46,7 @@ public class Issue {
         List<Long> total = issueLiveMap.entrySet()
                 .parallelStream().map(s -> s.getValue().close - s.getValue().open).collect(Collectors.toList());
         Long result = total.parallelStream().reduce((s1, s2) -> s1 + s2).get();
-        return (long) Math.ceil((double) result.doubleValue() / (double) size);
-
-
+        return (long) Math.ceil(result.doubleValue() / (double) size);
     }
 
 }
